Modifiche di movieRepository e MovieService + inizializzazione di MovieController

+++ backend/springboot/src/main/java/com/AssignmentTWEB/springboot/model/Movie.java
@@ -18,13 +18,13 @@ public class Movie {
     @Column(columnDefinition = "TEXT") //TEXT type
     private String tagline;
 
-    private double minute;
+    private float minute;
 
-    private double rating;
+    private float rating;
 
     public Movie(){}
 
-    public Movie(int id_movie, String name, String description, String tagline, double minute, double rating) {
+    public Movie(int id_movie, String name, String description, String tagline, float minute, float rating) {
         this.id_movie = id_movie;
         this.name = name;
         this.description = description;
@@ -46,8 +46,8 @@ public class Movie {
     public void setTagline(String tagline){this.tagline = tagline;}
 
     public double getMinute(){return minute;}
-    public void setMinute(double minute){this.minute = minute;}
+    public void setMinute(float minute){this.minute = minute;}
 
     public double getRating(){return rating;}
-    public void setRating(double rating){this.rating = rating;}
+    public void setRating(float rating){this.rating = rating;}
 }
